move server to module and tidy up logging
