+++ api/package.json
@@ -11,9 +11,16 @@
         {
             "name": "lilcoltsmith",
             "email": "lilcoltsmith@gmail.com"
+        },
+        {
+            "name": "arpanchhetri",
+            "email": "arpan.chhetri@hotmail.com"
         }
     ],
     "license": "MIT",
+    "author": {
+        "name": ""
+    },
     "dependencies": {
         "express": "^4.17.1",
         "moment": "^2.29.1",
