+++ api/src/loaders/index.js
@@ -5,7 +5,19 @@
  *  *  */
 
 const Logger = require('../modules/logger/index');
+const Server = require('../modules/server/express');
+
+// for logging the loaders
+const logger = new Logger('App');
+
+function logloader(){
+      //Log that loaders are loaded
+      logger.info('Loaders loaded');
+    }
+
 
 module.exports = {
-  Logger: Logger
+  Logger: Logger,
+  Server: Server,
+  Loader: logloader()
 }
